django_chat/requirements.txt
django
channels
daphne
requests
python-dotenv
google-genai
dj-database-url
psycopg2-binary
channels_redis
pypdf
sentence-transformers
django-allauth